+++ racetrack_commons/racetrack_commons/entities/dto.py
@@ -44,8 +44,10 @@ class JobDto(BaseModel):
     infrastructure_target: str | None = None
     # internal hostnames of the job replicas (eg. pods)
     replica_internal_names: list[str] = []
-    # exact name and version of a job type used to built this job
+    # exact name and version of a job type used to build this job
     job_type_version: str = ''
+    # Statistics of a Job seen by infrastructure
+    infrastructure_stats: dict[str, Any] = {}
 
     def __str__(self):
         return f'{self.name} v{self.version}'